updated delete across desktop and mobile, also fixed mobile tag filtering
